Committing changes such that all submetrics within a Timer do not have to be mandatorily present. Converting all of them to their boxed type. In serialize timer, only submetrics that are not null are added

+++ blueflood-core/src/integration-test/java/com/rackspacecloud/blueflood/service/PreaggregatedMetricsIntegrationTest.java
@@ -57,7 +57,7 @@ public class PreaggregatedMetricsIntegrationTest extends IntegrationTestBase {
             .withVariance(103d)
             .withMinValue(104)
             .withMaxValue(105)
-            .withCount(106);
+            .withCount((long)106);
         simple.setPercentile("98th", 107);
         simple.setPercentile("99th", 110);
     }
